refactor: create `renderNunjucksWithFrontmatter` function
To make unit testing easier, I've wrapped the code to handle Frontmatter with a function.
